add: missing input fields for asset form
