+++ src/components/assets/AssetList.tsx
@@ -1,60 +1,250 @@
-import { Button, buttonVariants } from "@/components/ui/button";
+import { Button } from "@/components/ui/button";
+import { Input } from "@/components/ui/input";
+import { PaginationComponent } from "@/components/ui/PaginationComponent";
+import { Skeleton } from "@/components/ui/skeleton";
+import {
+  Table,
+  TableBody,
+  TableCell,
+  TableHead,
+  TableHeader,
+  TableRow,
+} from "@/components/ui/table";
+import { Eye, EyeOff, Search, Settings, Trash2 } from "lucide-react";
+import { useState } from "react";
+
 import { useAssets } from "@/hooks/use-assets";
-import { History, Settings } from "lucide-react";
-import { Link } from "react-router-dom";
+import { Asset } from "@/types/asset";
+import { AssetDetailsDialog } from "./AssetDetailsDialog";
+import { AssetFormDialog } from "./AssetFormDialog";
 
 export const AssetList = () => {
-  const { useAssetsQuery } = useAssets();
+  const [isViewDetailsDialogOpen, setIsViewDetailsDialogOpen] = useState(false);
+  const [selectedItem, setSelectedItem] = useState<Asset | null>(null);
+  const [isEditDialogOpen, setIsEditDialogOpen] = useState(false);
+
+  const { useAssetsQuery, useAssetTypesQuery } = useAssets();
+
   const { data: assets, isLoading, error } = useAssetsQuery();
+
+  const [filter, setFilter] = useState("");
+  const [typeFilter, setTypeFilter] = useState<number | string>("");
+
+  const [currentPage, setCurrentPage] = useState(1);
+  const itemsPerPage = 10;
+
+  const { data: assetTypes = [], isLoading: isLoadingAssetType } =
+    useAssetTypesQuery();
 
   if (error) {
     console.error("Error fetching assets:", error);
     return <div>Error fetching assets</div>;
   }
 
-  if (isLoading) {
-    return <div>Loading assets...</div>;
+  if (isLoading || isLoadingAssetType) {
+    return (
+      <div className="space-y-4">
+        <Skeleton className="h-10 w-full" />
+        <Skeleton className="h-10 w-full" />
+        <Skeleton className="h-10 w-full" />
+      </div>
+    );
   }
+
+  const filteredAssets = (assets ?? [])
+    .filter((asset) => asset.name.toLowerCase().includes(filter.toLowerCase()))
+    .filter((asset) => (typeFilter ? asset.type.id === typeFilter : true));
+
+  const paginatedSites = filteredAssets.slice(
+    (currentPage - 1) * itemsPerPage,
+    currentPage * itemsPerPage
+  );
+
+  const totalPages = Math.ceil(filteredAssets.length / itemsPerPage);
+  const startItem = (currentPage - 1) * itemsPerPage + 1;
+  const endItem = Math.min(currentPage * itemsPerPage, filteredAssets.length);
 
   return (
     <div className="space-y-4">
-      <div className="rounded-md border p-4">
-        <h1 className="text-2xl font-bold mb-4">Asset List</h1>
-        <div className="flex flex-col gap-2">
-          {assets &&
-            assets.map((asset) => {
-              return (
-                <div
-                  key={asset.id}
-                  className="flex justify-between rounded-sm border bg-slate-50 p-2"
-                >
-                  <div className="flex flex-col p-1">
-                    <h2 className="text-md font-semibold">{asset.name}</h2>
-                    <p className="text-xs text-muted-foreground">
-                      {asset.type.name}
-                    </p>
-                  </div>
-                  <div className="flex space-x-2 p-1">
-                    <Button variant="outline" size="icon">
-                      <History className="h-4 w-4" />
-                    </Button>
-                    <Link
-                      className={buttonVariants({
-                        variant: "outline",
-                        size: "icon",
-                      })}
-                      to={{
-                        pathname: "/asset/detail/" + asset.id,
-                      }}
-                    >
-                      <Settings className="h-4 w-4" />
-                    </Link>
-                  </div>
-                </div>
-              );
-            })}
+      <div className="relative mb-4 flex space-x-4">
+        <Input
+          placeholder="Search asset.."
+          value={filter}
+          onChange={(e) => setFilter(e.target.value)}
+          className="pl-10"
+        />
+        <Search className="absolute top-1/2 transform -translate-y-1/2 h-5 w-5 text-gray-400" />
+        <div className="relative">
+          <select
+            value={typeFilter}
+            onChange={(e) =>
+              setTypeFilter(e.target.value ? Number(e.target.value) : "")
+            }
+            className="h-10 rounded-md border border-input bg-background px-3 py-2 text-sm focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2"
+          >
+            <option value="">All</option>
+            {assetTypes.map((type) => (
+              <option key={type.id} value={type.id}>
+                {type.name}
+              </option>
+            ))}
+          </select>
         </div>
       </div>
+      <div className="rounded-md border">
+        {isLoading ? (
+          <Table>
+            <TableHeader>
+              <TableRow>
+                <TableHead>No.</TableHead>
+                <TableHead>Item Name</TableHead>
+                <TableHead>Type</TableHead>
+                <TableHead>Quantity</TableHead>
+                <TableHead>Nadi Center</TableHead>
+                <TableHead>Request Date</TableHead>
+                <TableHead>Status</TableHead>
+                <TableHead>Actions</TableHead>
+              </TableRow>
+            </TableHeader>
+            <TableBody>
+              {Array.from({ length: 4 }).map((_, index) => (
+                <TableRow key={index}>
+                  <TableCell>
+                    <Skeleton className="h-4 w-6" />
+                  </TableCell>
+                  <TableCell>
+                    <Skeleton className="h-4 w-32" />
+                  </TableCell>
+                  <TableCell>
+                    <Skeleton className="h-4 w-24" />
+                  </TableCell>
+                  <TableCell>
+                    <Skeleton className="h-4 w-20" />
+                  </TableCell>
+                  <TableCell>
+                    <Skeleton className="h-4 w-24" />
+                  </TableCell>
+                  <TableCell>
+                    <Skeleton className="h-4 w-20" />
+                  </TableCell>
+                  <TableCell>
+                    <Skeleton className="h-4 w-20" />
+                  </TableCell>
+                  <TableCell>
+                    <div className="flex space-x-2">
+                      <Skeleton className="h-8 w-8 rounded-full" />
+                      <Skeleton className="h-8 w-8 rounded-full" />
+                    </div>
+                  </TableCell>
+                </TableRow>
+              ))}
+            </TableBody>
+          </Table>
+        ) : (
+          <Table>
+            <TableHeader>
+              <TableRow>
+                <TableHead>No.</TableHead>
+                <TableHead>Item Name</TableHead>
+                <TableHead>Type</TableHead>
+                <TableHead>Quantity</TableHead>
+                <TableHead>Nadi Center</TableHead>
+                <TableHead>Request Date</TableHead>
+                <TableHead>Status</TableHead>
+                <TableHead>Actions</TableHead>
+              </TableRow>
+            </TableHeader>
+            <TableBody>
+              {paginatedSites &&
+                paginatedSites.map((asset, index) => {
+                  const requestDate = asset.created_at
+                    ? asset.created_at.split("T")[0]
+                    : "";
+
+                  return (
+                    <TableRow key={asset.id}>
+                      <TableCell>
+                        {(currentPage - 1) * itemsPerPage + index + 1}
+                      </TableCell>
+                      <TableCell>{asset?.name || ""}</TableCell>
+                      <TableCell>{asset?.type.name || ""}</TableCell>
+                      <TableCell>{asset?.qty_unit || ""}</TableCell>
+                      <TableCell>{asset?.site.sitename || ""}</TableCell>
+                      <TableCell>{requestDate || ""}</TableCell>
+                      <TableCell>
+                        {asset?.is_active ? "Active" : "Inactive"}
+                      </TableCell>
+                      <TableCell>
+                        <Button
+                          variant="outline"
+                          size="icon"
+                          onClick={() => {}}
+                        >
+                          {asset.is_active ? (
+                            <Eye className="h-4 w-4" />
+                          ) : (
+                            <EyeOff className="h-4 w-4" />
+                          )}
+                        </Button>
+                        <Button
+                          variant="outline"
+                          size="icon"
+                          onClick={() => {
+                            setIsEditDialogOpen(true);
+                            setSelectedItem(asset);
+                          }}
+                        >
+                          <Settings className="h-4 w-4" />
+                        </Button>
+                        <Button
+                          variant="outline"
+                          size="icon"
+                          className="text-destructive"
+                          onClick={() => {}}
+                        >
+                          <Trash2 className="h-4 w-4" />
+                        </Button>
+                        <Button
+                          variant="outline"
+                          size="icon"
+                          onClick={() => {
+                            setIsViewDetailsDialogOpen(true);
+                            setSelectedItem(asset);
+                          }}
+                        >
+                          <Search className="h-4 w-4" />
+                        </Button>
+                      </TableCell>
+                    </TableRow>
+                  );
+                })}
+            </TableBody>
+          </Table>
+        )}
+      </div>
+      {totalPages > 1 && (
+        <PaginationComponent
+          currentPage={currentPage}
+          totalPages={totalPages}
+          onPageChange={setCurrentPage}
+          startItem={startItem}
+          endItem={endItem}
+          totalItems={filteredAssets.length}
+        />
+      )}
+
+      <AssetDetailsDialog
+        open={isViewDetailsDialogOpen}
+        onOpenChange={setIsViewDetailsDialogOpen}
+        asset={selectedItem}
+      />
+      {isEditDialogOpen && (
+        <AssetFormDialog
+          open={isEditDialogOpen}
+          onOpenChange={setIsEditDialogOpen}
+          asset={selectedItem}
+        />
+      )}
     </div>
   );
 };
